+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 1.2
 ---
 - Fix `preparable_query()` regular expression incorrect handling of numbers within database identifiers.
+- Make `autovacuum()` support older PostgreSQL versions all the way down to 9.3.
 
 1.1
 ---